.history/services/document-processor/requirements_20251011120646.txt
# FastAPI and web framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication and security
passlib[bcrypt]==1.7.4

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
python-magic==0.4.27
Pillow==10.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
mistune==3.0.2
lxml==4.9.3

# Text processing and NLP
spacy==3.7.2
nltk==3.8.1
langdetect==1.0.9
tiktoken==0.5.2

# Embeddings
openai==1.3.7
sentence-transformers==2.2.2

# Vector storage
qdrant-client==1.7.0

# Background jobs
dramatiq[redis]==1.15.0

# Retry logic
tenacity==8.2.3

# Object storage
minio==7.2.0

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.7.1
